feat: add toasts to AvatarProfileInput

+++ src/components/containers/AvatarProfileInput/AvatarProfileInput.tsx
@@ -1,5 +1,6 @@
 import {ChangeEvent, DragEventHandler, FC, useState} from "react";
 import {useParams} from "react-router-dom";
+import toast from "react-hot-toast";
 import Avatar from "@mui/material/Avatar";
 import Box from "@mui/material/Box";
 import IconButton from "@mui/material/IconButton";
@@ -12,6 +13,7 @@ import {convertAvatarToRequestData} from "@/utils/convertAvatarToRequestData";
 import {useUploadAvatar} from "@/graphql/hooks/useUploadAvatar";
 import {AvatarData} from "@/models/AvatarData.type";
 import {useDeleteAvatar} from "@/graphql/hooks/useDeleteAvatar";
+import {AuthInfoService} from "@/services/AuthInfoService";
 
 interface AvatarProfileInputProps {
   avatarPath: string;
@@ -26,26 +28,48 @@ export const AvatarProfileInput: FC<AvatarProfileInputProps> = ({
   const {uploadAvatar} = useUploadAvatar();
   const {deleteAvatar} = useDeleteAvatar();
   const [dragActive, setDragActive] = useState<boolean>(false);
+  const authorizedId = AuthInfoService.getAuthInfo().id;
 
   const uploadAvatarData = async (avatarData: AvatarData) => {
-    await uploadAvatar(id!, avatarData);
-    onLoadUserInfo();
+    const authorizedId = AuthInfoService.getAuthInfo().id;
+    if (id === authorizedId) {
+      await uploadAvatar(id!, avatarData);
+      onLoadUserInfo();
+    }
+    if (id !== authorizedId) {
+      toast("Cannot update photo of other profiles");
+    }
+  };
+
+  const validateFileAndUpload = (file: File | undefined) => {
+    if (id !== authorizedId) {
+      toast("cannot update photo of other profiles");
+      return;
+    }
+    if (file && id && file.size > 4000000) {
+      toast("File more than 0.5MB");
+      return;
+    }
+    if (file && id && file.size <= 4000000) {
+      convertAvatarToRequestData(file, uploadAvatarData);
+    }
   };
 
   const deleteAvatarHandler = async () => {
-    await deleteAvatar(id!);
-    onLoadUserInfo();
+    if (id === authorizedId) {
+      await deleteAvatar(id!);
+      onLoadUserInfo();
+    }
+    if (id !== authorizedId) {
+      toast("Сannot delete photo of other profiles");
+    }
   };
 
   const changeHandler = async (e: ChangeEvent<HTMLInputElement>) => {
     e.preventDefault();
+
     const file = e.target.files?.[0];
-    if (file && id && file.size <= 4000000) {
-      convertAvatarToRequestData(file, uploadAvatarData);
-    }
-    if (file && id && file.size > 4000000) {
-      console.log("File more than 0,5MB");
-    }
+    validateFileAndUpload(file);
   };
 
   const dragEnterHandler: DragEventHandler<HTMLFormElement> = e => {
@@ -70,11 +94,11 @@ export const AvatarProfileInput: FC<AvatarProfileInputProps> = ({
     e.preventDefault();
     e.stopPropagation();
     setDragActive(false);
-    if (e.dataTransfer.files[0]) {
-      const file = e.dataTransfer.files[0];
-      convertAvatarToRequestData(file, uploadAvatarData);
-    }
+
+    const file = e.dataTransfer.files[0];
+    validateFileAndUpload(file);
   };
+
   return (
     <form
       onDrop={dropHandler}
